Implement CreateDealModal component and integrate it into LeadPage and DealDashboard for deal creation functionality
